Move apiserver-proxy tests
