--- tests/apiserver-proxy-tests/README.md
@@ -1,7 +1,0 @@
-# Apiserver-proxy-tests
-
-## Details
-- Contains the `Dockerfile` for the image used in Kyma apiserver-proxy tests.
-- Contains the `fetch-token` application used for fetching authentication token from Dex.
-- Contains the `test.sh` script that runs tests for the chart.
-
