+++ tests/integration/apiserver-proxy/Dockerfile
