+++ tests/integration/apiserver-proxy/Gopkg.toml
