+++ tests/integration/apiserver-proxy/fetch-token/internal/auth.go
