+++ tests/integration/apiserver-proxy/fetch-token/internal/idtokenprovider.go
